refactor(database): decompose the discovery repository god-file by entity (WS-D)
internal/database/repository_discovery.go was a 1157-line god-file holding the
persistence layer for every discovery entity in one file. Split by entity into
role files within the same package (pure move — no API, signature or logic
change; verified by an identical top-level declaration set and an identical
sorted code-line multiset against the original):

- repository_discovery.go (66) — DiscoveryRepository struct, shared consts/error
                                  sentinels, and the shared OUI/time helpers
- wifinetwork.go (204)          — WiFiNetworkDB + CRUD/scan
- wifiaccesspoint.go (249)      — WiFiAccessPointDB + CRUD/scan
- networkproblem.go (362)       — NetworkProblemDB/ProblemSummaryDB + CRUD/scan
- ouivendor.go (104)            — OUIVendorDB + CRUD + MAC vendor lookup
- channelutilization.go (136)   — ChannelUtilizationDB + CRUD
- wifistats.go (88)             — GetWiFiStats + WiFiStatsDB

The shared DiscoveryRepository struct, error sentinels and consts stay in
repository_discovery.go (package-scoped, used across the entity files). Largest
resulting file is 362 LOC, clearing the file-size threshold and removing
repository_discovery.go as a STRICT-gate blocker for WS-Z.

Advances the Architecture Completion Plan WS-D (god-file decomposition).
